Fix GCC sign-conversion error in granule validation
GCC's -Wsign-conversion flags the int64_t += size_t accumulation in
validate_granule_position(): the unsigned operand forces the addition
to unsigned and the store back to signed, both of which it reports as
possible sign changes. Clang stays quiet, so it only broke the GCC CI
job once this branch made the wrapper warnings fatal.

The accumulator is guarded non-negative and the per-packet sample count
is tiny, so the conversion is safe. Cast decoded_samples to int64_t to
make that explicit and keep the addition signed.

+++ src/ogg_opus_decoder.cpp
@@ -135,7 +135,7 @@ OggOpusResult OggOpusDecoder::validate_granule_position(int64_t granule_pos, siz
         }
 
         if (first_audio_page_samples_ >= 0) {
-            first_audio_page_samples_ += decoded_samples;
+            first_audio_page_samples_ += static_cast<int64_t>(decoded_samples);
 
             if (is_last_on_page) {
                 if (!is_eos && granule_pos < first_audio_page_samples_) {
